+++ src/refresh/Tile.ts
@@ -15,7 +15,7 @@ class Tile {
   size: number;
   gl: WebGL2RenderingContext | WebGLRenderingContext;
   texture: WebGLTexture | undefined;
-  bbox: Float64Array; // 相对于画布的坐标，没有旋转所以只算2个顶点即可
+  bbox: Float32Array; // 相对于画布的坐标，没有旋转所以只算2个顶点即可
   x1: number;
   y1: number;
   x2: number;
@@ -36,7 +36,7 @@ class Tile {
     this.y = y;
     this.size = size;
     this.gl = gl;
-    this.bbox = new Float64Array([0, 0, 0, 0]);
+    this.bbox = new Float32Array([0, 0, 0, 0]);
     this.x1 = this.y1 = -1;
     this.x2 = this.y2 = 1;
   }
